furniture-website/requirements.txt
Flask
Jinja2
Werkzeug
click
gunicorn
python-dotenv